% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{boot}
\alias{boot}
\title{Nonparametric Bootstrap Estimation for Treatment Effects}
\usage{
boot(
  Z,
  X,
  S,
  Y,
  n_boot = 500,
  n_divisions = 100,
  copula_type = "gaussian",
  rho = 0,
  weighting_function_vectorized = identity_weighting_function_vectorized,
  g_function_vectorized = identity_g_function_vectorized
)
}
\arguments{
\item{Z}{A numeric vector indicating the treatment assignment (1 for treatment,
0 for control).}

\item{X}{A matrix of covariates.}

\item{S}{A numeric vector representing the post-treatment variable.}

\item{Y}{A numeric vector of outcomes.}

\item{n_boot}{Integer specifying the number of bootstrap iterations.
Default is 500.}

\item{n_divisions}{Integer specifying the number of divisions for integration.
Default is 100.}

\item{copula_type}{A string indicating the type of copula to be used.
Default is 'gaussian'.}

\item{rho}{A numeric parameter for the copula, the correlation coefficient, default is 0.}

\item{weighting_function_vectorized}{A vectorized weighting function used in
the calculation of estimators. Default is `identity_weighting_function_vectorized`.}

\item{g_function_vectorized}{A vectorized g-function used in the calculation of
estimators. Default is `identity_g_function_vectorized`.}
}
\value{
A list with the following elements:
\item{point_est}{The point estimates for projection coefficients,
as returned by the `point_estimator` function.}
\item{boot_est}{A matrix of bootstrap estimates across the `n_boot` iterations.}
\item{res}{A matrix with two rows: the first row contains the point estimates,
and the second row contains the bootstrap standard errors.}
}
\description{
This function performs nonparametric bootstrap to estimate the standard errors
of projection coefficients of principal causal effect estimators. The bootstrap samples are drawn with replacement
from the data, and for each sample, the `point_estimator` function is used to
estimate the projection coefficients. The function returns both the point estimates and
the bootstrap standard errors.
}
\examples{
# Example usage
set.seed(42)
Z <- sample(0:1, 100, replace = TRUE)
X <- matrix(rnorm(100*3), ncol = 3)
S <- rnorm(100)
Y <- rnorm(100)
boot_result <- boot(Z, X, S, Y, n_boot=5)
print(boot_result$res)

}
