% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copula_utils.R
\name{gaussian_copula_gradient}
\alias{gaussian_copula_gradient}
\title{Gradient of the Gaussian Copula}
\usage{
gaussian_copula_gradient(u, v, rho)
}
\arguments{
\item{u}{A numeric vector of uniform random variables representing the first
set of inputs. The values must be in the range (0, 1).}

\item{v}{A numeric vector of uniform random variables representing the second
set of inputs. The values must be in the range (0, 1).}

\item{rho}{A numeric scalar between -1 and 1, representing the correlation
parameter of the Gaussian copula.}
}
\value{
A list containing two components:
\itemize{
  \item \code{cu}: A numeric vector of the gradient of the Gaussian copula with respect to `u`.
  \item \code{cv}: A numeric vector of the gradient of the Gaussian copula with respect to `v`.
}
}
\description{
This function computes the gradients of the Gaussian copula with respect to
both components `u` and `v`.
}
\details{
The gradient of the Gaussian copula is calculated with respect to
both `u` and `v`. For each element of `u` and `v`, the function computes the
partial derivatives of the Gaussian copula with respect to the first and
second inputs, respectively.

The gradients are computed as follows: \deqn{c_u = \frac{\partial
C(u,v)}{\partial u} = \frac{C(u,v) \cdot \left( \rho y_v - \rho^2 x_u
\right)}{(1-\rho^2) dnorm(x_u)}} \deqn{c_v = \frac{\partial C(u,v)}{\partial
v} = \frac{C(u,v) \cdot \left( \rho x_u - \rho^2 y_v \right)}{(1-\rho^2)
dnorm(y_v)}} where \eqn{C(u, v)} is the Gaussian copula, and \eqn{x_{u}} and
\eqn{y_{v}} are the quantile transformations of `u` and `v`, respectively.
}
\examples{
# Example: Compute the gradient of the Gaussian copula with correlation 0.5
u <- runif(10)
v <- runif(10)
rho <- 0.5
gradients <- gaussian_copula_gradient(u, v, rho)

# Display the gradient components
gradients$cu
gradients$cv

}
\seealso{
\code{\link{gaussian_copula_function}} for computing the Gaussian
  copula.
}
