% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{identity_g_function_vectorized}
\alias{identity_g_function_vectorized}
\title{Vectorized G Function}
\usage{
identity_g_function_vectorized(s1, s0)
}
\arguments{
\item{s1}{A numeric vector of length `n_s1` representing the first set of
input values.}

\item{s0}{A numeric vector of length `n_s0` representing the second set of
input values.}
}
\value{
A three-dimensional numeric array of dimensions `n_s1 x n_s0 x 3`,
  containing:
  \itemize{
    \item The first slice of the array, corresponding to `s1` repeated for each element of `s0`.
    \item The second slice of the array, corresponding to `s0` repeated for each element of `s1`.
    \item A third slice filled with ones, representing a constant dimension.
  }
}
\description{
This function constructs a three-dimensional array by combining three
matrices.
}
\details{
The function generates a three-dimensional array by combining three
components: 1. A matrix where each element of `s1` is repeated for each value
in `s0`. 2. A matrix where each element of `s0` is repeated for each value in
`s1`. 3. A matrix filled with ones, serving as a constant dimension.
}
\examples{
# Example: Create a 3D array combining two vectors
s1 <- 1:3
s0 <- c(0.1, 0.5, 0.9)
result <- identity_g_function_vectorized(s1, s0)

# Display the resulting 3D array
print(result)

}
