% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eif_estimator_utils.R
\name{int_s0}
\alias{int_s0}
\title{Compute the Integral over s0}
\usage{
int_s0(
  s1,
  psp_s0,
  psp_s1_cdf,
  psp_s0_cdf,
  Z,
  Tp,
  S,
  copula_type = "independent",
  rho = 0
)
}
\arguments{
\item{s1}{A numeric vector of size \eqn{n_{s1}} representing the first set of
values.}

\item{psp_s0}{A matrix of size (\eqn{n_{s0}}, n), representing the
probabilities associated with the second set of values s0.}

\item{psp_s1_cdf}{A matrix of size (\eqn{n_{s1}}, n) representing the
cumulative distribution function (CDF) of s1.}

\item{psp_s0_cdf}{A matrix of size (\eqn{n_{s0}}, n) representing the
cumulative distribution function (CDF) of s0.}

\item{Z}{A numeric vector of size n representing the treatment assignment indicator.}

\item{Tp}{A numeric value representing the fitted treatment probability.}

\item{S}{A numeric vector of size n representing the post-treatment variable.}

\item{copula_type}{A character string indicating the type of copula to be
used. Defaults to "independent". Other options are "gaussian" and "fgm".}

\item{rho}{A numeric value between -1 and 1 representing the correlation
coefficient for copula models. Default is 0 (for independence).}
}
\value{
A matrix of size (\eqn{n_{s0}}, n).
}
\description{
Compute the Integral over s0
}
\examples{
# Example usage of the function
s1 <- rnorm(20)  # First set of values (e.g., target variable)
s0 <- rnorm(5)   # Second set of values (e.g., reference variable)
psp_s1 <- matrix(runif(100), 20, 10)   # Probability values for s1
psp_s0 <- matrix(runif(50), 5, 10)   # Probability values for s0
psp_s1_cdf <- trapz_cdf(s1, psp_s1)  # CDF of s1
psp_s0_cdf <- trapz_cdf(s0, psp_s0)  # CDF of s0
Z <- sample(0:1, 10, replace = TRUE)  # Treatment assignment indicator
Tp <- 0.7  # Event probability
S <- 1:10  # Sequential index

# Call the function
result <- int_s0(s1, psp_s0, psp_s1_cdf, psp_s0_cdf, Z, Tp, S)
print(dim(result))

}
