% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_utils.R
\name{matrix_multiply_with_expansion}
\alias{matrix_multiply_with_expansion}
\title{Matrix Multiply with Expansion}
\usage{
matrix_multiply_with_expansion(U, V, expansion_type = "both")
}
\arguments{
\item{U}{A numeric matrix of size (\eqn{n_{u}}, n) or of size (\eqn{n_{u}},
\eqn{n_{v}}, n). When `expansion_type = "U" or "both"`, this matrix will be
expanded.}

\item{V}{A numeric matrix of size (\eqn{n_{v}}, n) or of size (\eqn{n_{u}},
\eqn{n_{v}}, n). When `expansion_type = "V" or "both"`, this matrix will be
expanded.}

\item{expansion_type}{A character string specifying the type of expansion.
One of:
\itemize{
  \item "U": Expand the dimensions of `U` to match the number of rows in `V`.
  \item "V": Expand the dimensions of `V` to match the number of rows in `U`.
  \item "both" (default): Expand both `U` and `V` to match the respective dimensions.
}}
}
\value{
A numeric array of size (\eqn{n_{u}}, \eqn{n_{v}}, n), where
  \eqn{n_{u}} and \eqn{n_{v}} are the dimensions of `U` and `V` respectively.
}
\description{
Perform element-wise multiplication of two matrices `U` and `V`, with the
option to expand dimensions of either matrix in a specified way. The function
supports three expansion types: expanding `U`, expanding `V`, or expanding
both matrices to match their respective dimensions.
}
\examples{
# Example 1: Expanding "both"
U <- matrix(1:6, nrow = 2, ncol = 3)
V <- matrix(7:18, nrow = 4, ncol = 3)
result <- matrix_multiply_with_expansion(U, V, expansion_type = "both")

}
