% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{principal_score_predict_vectorized}
\alias{principal_score_predict_vectorized}
\title{Principal Score Prediction (Vectorized)}
\usage{
principal_score_predict_vectorized(bw, S, X)
}
\arguments{
\item{bw}{A bandwidth object generated from the `npcdensbw` function. This
contains the bandwidth parameter for kernel density estimation.}

\item{S}{A numeric vector of length \eqn{n_s}, containing the target variable
values for which predictions will be made.}

\item{X}{A numeric matrix with dimensions (n, dims(X)), or data frame
containing the predictor variables for which principal score predictions
are to be made.}
}
\value{
A numeric matrix of predicted values with dimensions (\eqn{n_s},  n),
  where each row corresponds to predictions for the respective values in S.
}
\description{
This function performs vectorized prediction of principal scores based on a
given bandwidth object and predictor variables. It uses the kernel density
estimation function `npcdens` from the `np` package to predict the target
variable given a set of predictors.
}
\details{
This function leverages the `npcdens` function from the `np` package
to perform kernel density estimation using the bandwidth object provided in
`bw`. The target variable `S` is predicted based on the predictors `X` using
the nonparametric kernel density estimate, where each row of the resulting
output corresponds to a predicted value for a corresponding target value in
`S`. The computation is vectorized for efficiency.
}
\examples{
# Example: Compute principal score predictions
set.seed(42)
X <- matrix(rnorm(200), ncol=2)  # Predictor variables
S <- rnorm(100)  # Target variable

# Compute bandwidth using the npcdensbw function
bw <- np::npcdensbw(xdat=X, ydat=S)

# Compute predicted values for principal scores
predictions <- principal_score_predict_vectorized(bw, S, X)
print(predictions)

}
\seealso{
\code{\link[np]{npcdens}} for kernel density estimation and
  \code{\link[np]{npcdensbw}} for bandwidth selection.
}
