% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{wg_vectorized}
\alias{wg_vectorized}
\title{Vectorized \eqn{W * G} Computation}
\usage{
wg_vectorized(
  s1,
  s0,
  weighting_function_vectorized = identity_weighting_function_vectorized,
  g_function_vectorized = identity_g_function_vectorized
)
}
\arguments{
\item{s1}{A numeric vector of size \eqn{n_{s1}} representing the first set of
values.}

\item{s0}{A numeric vector of size \eqn{n_{s0}} representing the second set
of values.}

\item{weighting_function_vectorized}{A function that computes the weights
based on `s1` and `s0`. The default is
\code{identity_weighting_function_vectorized}.}

\item{g_function_vectorized}{A function that computes the generalized
transformation based on `s1` and `s0`. The default is
\code{identity_g_function_vectorized}.}
}
\value{
A numeric array of size (\eqn{n_{s1}, n_{s0}, dim_g}), where
  \eqn{dim_g} corresponds to the third dimension of the output from the
  `g_function_vectorized`.
}
\description{
This function computes a vectorized version of the weighting and
transformation between two sets of values `s1` and `s0`, using the specified
weighting function and transformation function.
}
\details{
The function performs a series of vectorized operations between the
two input vectors `s1` and `s0`, applying the specified weighting function
(`weighting_function_vectorized`) and the transformation function
(`g_function_vectorized`) to these values. The results are then combined in a
multi-dimensional array.
}
\examples{
# Example: Compute vectorized weighted transformation
s1 <- rnorm(10)  # First set of values (e.g., target variable)
s0 <- rnorm(5)   # Second set of values (e.g., reference variable)
result <- wg_vectorized(s1, s0)
print(dim(result))

}
